[Product] Feat: 예약 실행 스케줄러 구현

+++ product-service/src/main/java/com/bubbletea/product/domain/event/ProductPriceChangeScheduledEvent.java
@@ -4,10 +4,10 @@ import java.time.LocalDateTime;
 
 
 public record ProductPriceChangeScheduledEvent(
-        String productName,
-        LocalDateTime priceChangeDate,
-        int originalPrice,
-        int changedPrice
+    String productName,
+    LocalDateTime priceChangeDate,
+    long originalPrice,
+    long changedPrice
 ) implements ProductEvent {
 
     @Override
